app/ai/computer_vision.py
"""
Computer Vision Module for Image Analysis
Uses transfer learning with pre-trained models (ResNet, EfficientNet, or Mask R-CNN)
to estimate square footage and severity from uploaded images

NOTE: AI dependencies are optional. If not installed, returns mock data.
"""
from typing import Dict, Tuple, List
import logging

# Try to import AI dependencies (optional)
try:
    import torch
    import torchvision.transforms as transforms
    from torchvision.models import resnet50, ResNet50_Weights
    from PIL import Image
    import cv2
    import numpy as np
    AI_AVAILABLE = True
except ImportError:
    AI_AVAILABLE = False
    logger = logging.getLogger(__name__)
    logger.warning("AI dependencies not installed. Using mock image analysis.")

from app.models.job import SeverityLevel, JobType

logger = logging.getLogger(__name__)


class ImageAnalyzer:
    """
    Computer Vision analyzer for estimating work size and severity
    Uses transfer learning with ResNet50 as feature extractor
    """
    
    def __init__(self):
        """Initialize the model and transformations"""
        self.ai_enabled = AI_AVAILABLE
        
        if self.ai_enabled:
            # Load pre-trained ResNet50
            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
            logger.info(f"Using device: {self.device}")
            
            # Load pre-trained model
            self.model = resnet50(weights=ResNet50_Weights.IMAGENET1K_V2)
            self.model = self.model.to(self.device)
            self.model.eval()
            
            # Image preprocessing
            self.transform = transforms.Compose([
                transforms.Resize((224, 224)),
                transforms.ToTensor(),
                transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
            ])
        else:
            logger.warning("AI model not initialized - AI dependencies not available")
        
        # Scaling factors for square footage estimation (simplified heuristic)
        # In production, this would be replaced with a fine-tuned model
        self.sqft_scaling_factors = {
            JobType.SNOW_REMOVAL: 1.5,
            JobType.LAWN_CARE: 1.2,
            JobType.HANDYMAN: 1.0,
        }
    
    async def analyze_image(
        self, 
        image_path: str, 
        job_type: JobType
    ) -> Dict[str, any]:
        """
        Analyze an image to estimate square footage and severity
        
        Args:
            image_path: Path to the image file
            job_type: Type of job (snow removal, lawn care, etc.)
            
        Returns:
            Dictionary with analysis results
        """
        # Return mock data if AI not available
        if not self.ai_enabled:
            return self._mock_analysis(job_type)
        
        try:
            # Load and preprocess image
            image = Image.open(image_path).convert('RGB')
            image_tensor = self.transform(image).unsqueeze(0).to(self.device)
            
            # Extract features using ResNet
            with torch.no_grad():
                features = self.model(image_tensor)
            
            # Estimate square footage using heuristic
            # In production: Use a fine-tuned regression head
            estimated_sqft = self._estimate_square_footage(
                image_path, features, job_type
            )
            
            # Estimate severity
            severity, confidence = self._estimate_severity(
                image_path, features, job_type
            )
            
            # Detect key objects/features
            detected_objects = self._detect_objects(image_path, job_type)
            
            return {
                "estimated_square_footage": estimated_sqft,
                "severity": severity,
                "confidence": confidence,
                "detected_objects": detected_objects,
                "metadata": {
                    "image_dimensions": image.size,
                    "job_type": job_type.value
                }
            }
            
        except Exception as e:
            logger.error(f"Error analyzing image: {e}")
            raise
    
    def _estimate_square_footage(
        self, 
        image_path: str, 
        features: torch.Tensor,
        job_type: JobType
    ) -> float:
        """
        Estimate square footage from image features
        
        In production, this would be a fine-tuned regression model.
        For MVP, we use heuristics based on image dimensions and feature statistics.
        """
        # Load image for dimension analysis
        img = cv2.imread(image_path)
        height, width = img.shape[:2]
        
        # Simple heuristic: Use image area as proxy
        # Assuming standard smartphone camera (adjust calibration in production)
        pixel_area = height * width
        
        # Convert to square feet (rough approximation)
        # This would be learned from training data in production
        base_sqft = (pixel_area / 10000) * 100  # Simplified conversion
        
        # Apply job-type specific scaling
        scaling_factor = self.sqft_scaling_factors.get(job_type, 1.0)
        estimated_sqft = base_sqft * scaling_factor
        
        # Clamp to reasonable ranges
        estimated_sqft = max(50, min(5000, estimated_sqft))
        
        return round(estimated_sqft, 2)
    
    def _estimate_severity(
        self,
        image_path: str,
        features: torch.Tensor,
        job_type: JobType
    ) -> Tuple[SeverityLevel, float]:
        """
        Estimate severity level from image features
        
        For snow: light/moderate/heavy based on visible ground
        For lawn: grass height and density
        For handyman: damage extent
        """
        # Load image for analysis
        img = cv2.imread(image_path)
        
        if job_type == JobType.SNOW_REMOVAL:
            severity, confidence = self._analyze_snow_severity(img)
        elif job_type == JobType.LAWN_CARE:
            severity, confidence = self._analyze_lawn_severity(img)
        else:
            # Default for handyman/other jobs
            severity = SeverityLevel.MODERATE
            confidence = 0.6
        
        return severity, confidence
    
    def _analyze_snow_severity(self, img: np.ndarray) -> Tuple[SeverityLevel, float]:
        """Analyze snow coverage and depth"""
        # Convert to HSV for better snow detection
        hsv = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
        
        # Snow is typically high value (brightness) and low saturation
        # Simple thresholding for MVP
        lower_white = np.array([0, 0, 200])
        upper_white = np.array([180, 30, 255])
        
        mask = cv2.inRange(hsv, lower_white, upper_white)
        snow_coverage = np.sum(mask > 0) / mask.size
        
        # Classify based on coverage
        if snow_coverage > 0.8:
            return SeverityLevel.SEVERE, 0.85
        elif snow_coverage > 0.5:
            return SeverityLevel.HEAVY, 0.75
        elif snow_coverage > 0.3:
            return SeverityLevel.MODERATE, 0.70
        else:
            return SeverityLevel.LIGHT, 0.65
    
    def _analyze_lawn_severity(self, img: np.ndarray) -> Tuple[SeverityLevel, float]:
        """Analyze grass height and condition"""
        # Convert to HSV
        hsv = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
        
        # Detect green areas (grass)
        lower_green = np.array([35, 40, 40])
        upper_green = np.array([85, 255, 255])
        
        mask = cv2.inRange(hsv, lower_green, upper_green)
        green_coverage = np.sum(mask > 0) / mask.size
        
        # Analyze texture for grass height (simplified)
        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        laplacian = cv2.Laplacian(gray, cv2.CV_64F)
        texture_variance = laplacian.var()
        
        # Higher variance often indicates taller/denser grass
        if texture_variance > 1000:
            return SeverityLevel.HEAVY, 0.75
        elif texture_variance > 500:
            return SeverityLevel.MODERATE, 0.70
        else:
            return SeverityLevel.LIGHT, 0.65
    
    def _detect_objects(self, image_path: str, job_type: JobType) -> List[str]:
        """
        Detect relevant objects in the image
        
        In production: Use object detection model (YOLO, Mask R-CNN)
        For MVP: Return placeholder based on job type
        """
        # Placeholder - in production, use actual object detection
        object_categories = {
            JobType.SNOW_REMOVAL: ["driveway", "sidewalk", "vehicle"],
            JobType.LAWN_CARE: ["lawn", "grass", "fence", "trees"],
            JobType.HANDYMAN: ["door", "window", "wall", "fence"]
        }
        
        return object_categories.get(job_type, ["general"])
    
    def _mock_analysis(self, job_type: JobType) -> Dict[str, any]:
        """
        Return mock analysis when AI dependencies are not available
        """
        mock_data = {
            JobType.SNOW_REMOVAL: {
                "estimated_square_footage": 500.0,
                "severity": SeverityLevel.MODERATE,
                "confidence": 0.7,
                "detected_objects": ["driveway", "sidewalk"],
            },
            JobType.LAWN_CARE: {
                "estimated_square_footage": 750.0,
                "severity": SeverityLevel.LIGHT,
                "confidence": 0.7,
                "detected_objects": ["lawn", "grass"],
            },
            JobType.HANDYMAN: {
                "estimated_square_footage": 200.0,
                "severity": SeverityLevel.MODERATE,
                "confidence": 0.7,
                "detected_objects": ["general"],
            },
        }
        
        base_result = mock_data.get(job_type, {
            "estimated_square_footage": 400.0,
            "severity": SeverityLevel.MODERATE,
            "confidence": 0.7,
            "detected_objects": ["general"],
        })
        
        base_result["metadata"] = {
            "image_dimensions": (1920, 1080),
            "job_type": job_type.value,
            "mock_data": True
        }
        
        return base_result


# Singleton instance
_image_analyzer = None

def get_image_analyzer() -> ImageAnalyzer:
    """Get or create ImageAnalyzer singleton"""
    global _image_analyzer
    if _image_analyzer is None:
        _image_analyzer = ImageAnalyzer()
    return _image_analyzer
